Secure the application

+++ app/api/issues/[id]/route.ts
@@ -1,48 +1,61 @@
+import authOptions from "@/app/auth/authOptions";
 import { issueSchema } from "@/app/ValidationSchema";
 import { prisma } from "@/prisma/client";
+import { getServerSession } from "next-auth";
 import { NextRequest, NextResponse } from "next/server";
 
-export async function PATCH(request: NextRequest, {params}: {params: Promise<{id: string}>}) {
-    const {id} = await params;
-    
-    const body = await request.json();
-    const validation = issueSchema.safeParse(body);
+export async function PATCH(
+  request: NextRequest,
+  { params }: { params: Promise<{ id: string }> }
+) {
+  const session = await getServerSession(authOptions);
+  if (!session) return NextResponse.json({}, { status: 401 });
+  const { id } = await params;
 
-    if (!validation.success) {
-        return NextResponse.json(validation.error.format(), {status: 400});
-    }
+  const body = await request.json();
+  const validation = issueSchema.safeParse(body);
 
-    const issue = await prisma.issue.findUnique({
-        where: {id: parseInt(id)}
-    })
+  if (!validation.success) {
+    return NextResponse.json(validation.error.format(), { status: 400 });
+  }
 
-    if (!issue)
-        return NextResponse.json({ error: 'Invalid issue'}, {status: 404});
+  const issue = await prisma.issue.findUnique({
+    where: { id: parseInt(id) },
+  });
 
-    const updatedIssue = await prisma.issue.update({
-        where: {id: issue.id},
-        data: {
-            title: body.title,
-            description: body.description
-        }
-    })
+  if (!issue)
+    return NextResponse.json({ error: "Invalid issue" }, { status: 404 });
 
-    return NextResponse.json(updatedIssue);
+  const updatedIssue = await prisma.issue.update({
+    where: { id: issue.id },
+    data: {
+      title: body.title,
+      description: body.description,
+    },
+  });
+
+  return NextResponse.json(updatedIssue);
 }
 
-export async function DELETE(request: NextRequest, {params}: {params: Promise<{id: string}>}) {
-    const {id} = await params;
+export async function DELETE(
+  request: NextRequest,
+  { params }: { params: Promise<{ id: string }> }
+) {
+  const session = await getServerSession(authOptions);
+  if (!session) return NextResponse.json({}, { status: 401 });
 
-    const issue = await prisma.issue.findUnique({
-        where: {id: parseInt(id)}
-    })
+  const { id } = await params;
 
-    if (!issue)
-        return NextResponse.json({error: 'Invalid issue'}, {status: 404});
+  const issue = await prisma.issue.findUnique({
+    where: { id: parseInt(id) },
+  });
 
-    await prisma.issue.delete({
-        where: {id: issue.id}
-    })
+  if (!issue)
+    return NextResponse.json({ error: "Invalid issue" }, { status: 404 });
 
-    return NextResponse.json({});
+  await prisma.issue.delete({
+    where: { id: issue.id },
+  });
+
+  return NextResponse.json({});
 }
